address process <-> connection manager race
Because the process manager runs independently of the connection manager, very rarely a race will occur where we get the connection open event before the process manager has discovered the process.

Allow a 50ms buffer to avoid this race. If we time out we will proceed with creating the connection without the process. On following events, the connection manager will attempt to rediscover the process.