adds rfc compliant http1 parser (#132)
Our original goal  of the http1 parser was to offload as much as possible to the Go standard library. Unfortunately, this was rather difficult to do because the std lib http package is designed around being a consumer or producer of http messages. This means it does makes some changes to wire traffic that work for it. While most of these changes were fine, a few were problematic when it came to a out-of-band observability use case. It also became difficult to handle traffic that was unhealthy, which we still wanted for observability purposes but had to hack around to get pass validity checks within the standard library.

This PR introduces a new HTTP1.0 and HTTP1.1 parser. I designed it following the previous parsers high level streaming operations, but reimplementing the internals using small pieces of the Go http package (eg. Headers struct, and chunk reader implementation), but left the rest of it to my own devices. I focused pretty heavily on satisfying the [RFC 9112](https://www.rfc-editor.org/rfc/rfc9112.html#name-handling-incomplete-message) spec.