Enhance E2E Suite tooling (#137)
This brings in a suite of tooling that can be used to quickly run test scenarios against a matrix of configurations. It's still early but will hopefully give us the foundation to build out a set of long and short running e2e tests to prevent future regressions and keep an eye on perf numbers.